backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
mysql-connector-python
pydantic
pytest
requests
